+++ src/databricks/labs/ucx/source_code/linters/pyspark.py
@@ -11,9 +11,8 @@ from databricks.labs.ucx.source_code.base import (
     Fixer,
     Linter,
 )
-from databricks.labs.ucx.source_code.linters.imports import ASTLinter, TreeWalker
 from databricks.labs.ucx.source_code.queries import FromTable
-from databricks.labs.ucx.source_code.linters.python_ast import AstHelper
+from databricks.labs.ucx.source_code.linters.python_ast import AstHelper, ASTLinter, TreeWalker
 
 
 @dataclass
