+++ tests/unit/source_code/linters/test_pyspark.py
@@ -3,7 +3,7 @@ import pytest
 from astroid import Call, Const, Expr  # type: ignore
 
 from databricks.labs.ucx.source_code.base import Deprecation, CurrentSessionState
-from databricks.labs.ucx.source_code.linters.imports import ASTLinter
+from databricks.labs.ucx.source_code.linters.python_ast import ASTLinter
 from databricks.labs.ucx.source_code.linters.pyspark import AstHelper, TableNameMatcher, SparkSql
 from databricks.labs.ucx.source_code.queries import FromTable
