+++ tests/unit/source_code/linters/test_python_linter.py
@@ -5,7 +5,8 @@ import pytest
 from astroid import Attribute, Call, Const, Expr  # type: ignore
 from databricks.labs.ucx.source_code.graph import DependencyProblem
 
-from databricks.labs.ucx.source_code.linters.imports import ASTLinter, DbutilsLinter, TreeWalker
+from databricks.labs.ucx.source_code.linters.imports import DbutilsLinter
+from databricks.labs.ucx.source_code.linters.python_ast import ASTLinter, TreeWalker
 
 
 def test_linter_returns_empty_list_of_dbutils_notebook_run_calls():
